Colour windspeed. Respect transmission ratio influence on torque constant.
The generator is not ideal. The Kv could be much lower, the generator's
RPM is way too high currently, which can be compensated with a high
transmission ratio though.

Nevertheless for a push + lift drag rotor as planned, the setup is not optimal, because for the aerodynamic lift a TSR of 6 to 7 is recommended, while this may depend several factors in reality:

For example it may depend on the air flow laminar separation point on the profile from at a certain high velocity, thus this presents a limit to the upper end. Therefore a TSR for 6 is trageted for until deeper knowledge has emerged. Currently the TSR is set to 1 to deliver the required high generator RPM.

Nominal voltage operation currently at 10m/s.

The current or voltage seem to be calculated in a wrong way. It'd be expected that the nominal power is reached at the nominal voltage. Either the current formula must include the power factor or I've read the wrong parameters (probably I guessed with the power factor).